+++ packages/moonstone/CHANGELOG.md
@@ -11,6 +11,11 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Fixed
 
 - `moonstone/Expandable` and derivatives to use the new `ease-out-quart` animation timing function to better match the aesthetic of Enyo's Expandables
+- `moonstone/LabeledItem` to start marquee when hovering over in disabled state
+- `moonstone/Marquee` to correctly start when hovering on disabled spottable components
+- `moonstone/Marquee.MarqueeController` to not abort marquee when moving among components
+- `moonstone/Picker` marquee issues with disabled buttons or Picker
+- `moonstone/Panels` to prevent loss of spotlight issue when moving between panels
 
 ## [1.12.2] - 2017-11-15